Client/requirements.txt
opencv-python
armonik==3.17.1
grpcio>=1.63.0
numpy
black
